+++ src/domain/pricing/details/sections/prices-details/prices-table/index.tsx
