+++ src/domain/pricing/details/sections/prices-details/prices-table/use-columns.tsx
