+++ src/domain/pricing/details/sections/prices-details/utils.ts
